Fix some clang warnings

+++ src/lib/WP5PageFormatGroup.cpp
@@ -150,7 +150,7 @@ void WP5PageFormatGroup::_readContents(WPXInputStream *input, WPXEncryption *enc
 					}
 				}
 			}
-			input->seek(20 - (m_tabStops.size() / 2 ) - (m_tabStops.size() % 2), WPX_SEEK_CUR);
+			input->seek(20 - long(m_tabStops.size() / 2 ) - long(m_tabStops.size() % 2), WPX_SEEK_CUR);
 
 			if ((getSize() > 4) && (getSize() - 4 == 0x00D0))
 			{
